+++ changes.md
@@ -1,5 +1,10 @@
 # Recent changes
 
+## 1 July 2023
+- Moved `create new additional style file` to the dropdown menu
+- Removed `merge into master` - now it happens automatically when `Edit additional` is unchecked
+- Fixed a crashing bug when a style had no name
+
 ## 30 June 2023
 - Delete from master list removes from additional style file as well
 - Major refactoring of code - sorry if I broke things, but it's going to be a lot easier going forward.